space page is done

+++ backend/internal/app/member_service.go
@@ -1,6 +1,9 @@
 package app
 
-import domain "github.com/si1ent-he11/AuraFlow/internal/domain/entity"
+import (
+	domain "github.com/si1ent-he11/AuraFlow/internal/domain/entity"
+	"github.com/si1ent-he11/AuraFlow/internal/domain/types"
+)
 
 func (as appService) CreateSpaceMember(dto domain.CreateSpaceMemberByInviteDTO, userId int) error {
 	spaceId, err := as.db.UseInvite(dto.InviteId)
@@ -28,22 +31,74 @@ func (as appService) CreateSpaceMember(dto domain.CreateSpaceMemberByInviteDTO, 
 	})
 }
 
-func (as appService) GetRoleByUserId(dto domain.SpaceMemberDTO) (string, error) {
-	return as.db.GetRoleByUserId(dto)
+func (as appService) GetMemberByMemberId(memberId int, spaceId int) (domain.SpaceMember, error) {
+	return as.db.GetMemberByMemberId(memberId, spaceId)
+}
+
+func (as appService) GetMember(dto domain.SpaceMemberDTO) (domain.SpaceMember, error) {
+	return as.db.GetMember(dto)
+}
+
+func (as appService) MemberExists(dto domain.SpaceMemberDTO) (bool, error) {
+	return as.db.MemberExists(dto)
+}
+
+func (as appService) UpdateMemberUsername(dto domain.UpdateSpaceMemberDTO) error {
+	return as.db.UpdateMemberUsername(dto)
+}
+
+func (as appService) GetAllAdminsFromSpace(dto domain.SpaceIdDTO) ([]domain.SpaceMember, error) {
+	return as.db.GetAllAdminsFromSpace(dto)
 }
 
 func (as appService) GetMembersFromSpace(dto domain.SpaceIdDTO) ([]domain.SpaceMember, error) {
 	return as.db.GetAllMembersFromSpace(dto)
 }
 
-func (as appService) PromoteMember(dto domain.SpaceMemberDTO) error {
-	return as.PromoteMember(dto)
+func (as appService) PromoteMember(dto domain.SpaceMemberDTO, ownerId int) error {
+	isOwner, err := as.db.IsOwner(domain.SpaceMemberDTO{
+		SpaceId: dto.SpaceId,
+		UserId:  ownerId,
+	})
+	if err != nil {
+		return err
+	}
+
+	if !isOwner {
+		return types.ErrorForbidden
+	}
+
+	return as.db.PromoteMember(dto)
 }
 
-func (as appService) DemoteMember(dto domain.SpaceMemberDTO) error {
+func (as appService) DemoteMember(dto domain.SpaceMemberDTO, ownerId int) error {
+	isOwner, err := as.db.IsOwner(domain.SpaceMemberDTO{
+		SpaceId: dto.SpaceId,
+		UserId:  ownerId,
+	})
+	if err != nil {
+		return err
+	}
+
+	if !isOwner {
+		return types.ErrorForbidden
+	}
+
 	return as.db.DemoteMember(dto)
 }
 
-func (as appService) DeleteMember(dto domain.SpaceMemberDTO) error {
+func (as appService) DeleteMember(dto domain.SpaceMemberDTO, ownerId int) error {
+	isOwner, err := as.db.IsOwner(domain.SpaceMemberDTO{
+		SpaceId: dto.SpaceId,
+		UserId:  ownerId,
+	})
+	if err != nil {
+		return err
+	}
+
+	if !isOwner {
+		return types.ErrorForbidden
+	}
+
 	return as.db.DeleteMember(dto)
 }
